requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
exceptiongroup==1.2.2
flake8==7.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
mccabe==0.7.0
portalocker==2.10.1
pycodestyle==2.12.1
pydub==0.25.1
pyflakes==3.2.0
pyTelegramBotAPI==4.14.1
python-dotenv==0.20.0
# python-telegram-bot==21.7
pywin32==308
requests==2.32.3
sniffio==1.3.1
SpeechRecognition==3.10.1
typing_extensions==4.12.2
urllib3==2.2.3
